// backup_2026-02-24_161500/types.ts
export interface ItemAllocationConfig {
  mode: 'ALL' | 'NON_DRINKER' | 'DRINKER' | 'CUSTOM';
  customQtyNonDrinker?: number;
  customQtyDrinker?: number;
}